+++ packages/plugin-ice-ssr/src/index.ts
@@ -111,8 +111,6 @@ const plugin = async (api): Promise<void> => {
 
     // remove process fallback when target is node
     config.plugins.delete('ProvidePlugin');
-    // no need to generate the loadable-stats.json because it will not be used in anywhere
-    config.plugins.delete('@loadable/webpack-plugin');
 
     async function serverRender(res, req) {
       const htmlTemplate = fse.readFileSync(path.join(buildDir, 'index.html'), 'utf8');
@@ -213,12 +211,6 @@ const plugin = async (api): Promise<void> => {
     const minifiedHtml = minify(html, { collapseWhitespace: true, quoteCharacter: '\'' }).replace(/`/g, '&#x60;');
     const newBundle = bundle.replace(/'global.__ICE_SERVER_HTML_TEMPLATE__'/, `\`${minifiedHtml}\``);
     fse.writeFileSync(serverFilePath, newBundle, 'utf-8');
-
-    // @deprecated
-    // In the future, the loadable-stats.json file in the server dir will be removed because it will be used in anywhere.
-    // copy web loadable-stats.json to server directory in order to be compatible
-    const statsFileName = 'loadable-stats.json';
-    fse.copyFileSync(path.join(buildDir, statsFileName), path.join(serverDir, statsFileName));
   });
 };